fix: исправление ошибое ревью 2
